EnemyShip implementations

+++ OLCTest/Asteroid.cpp
@@ -2,7 +2,7 @@
 #include "MovingSpaceObject.h";
 #include "SpaceObject.h";
 
-Asteroid::Asteroid(double dimX, double dimY, olc::Sprite* sprite, double speed) :MovingSpaceObject(SpaceObjectConstants::SPACE_X - dimX, (rand() % (SpaceObjectConstants::SPACE_Y - dimY)), dimX, dimY, sprite, speed) {};
+Asteroid::Asteroid(int dimX, int dimY, olc::Sprite* sprite, double speed) :MovingSpaceObject(SpaceObjectConstants::SPACE_X - dimX, (rand() % (SpaceObjectConstants::SPACE_Y - dimY)), dimX, dimY, sprite, speed) {};
 
 void Asteroid::move(float elapsedTime) {
 	if (this->xPosition <= 0) {
